requirements.txt
Django==1.11.2
hurry.filesize==0.9